Another way to find the document to be annotated from sections file.
There is no way to know what document is going to be annotated in sections files. Before, the element annotated with the type bibo:AcademicArticle was retrieved. If there were more than one, then the first one was selected. This prevent the annotation of other documents annotated with other type. Therefore, only documents annotated with bibo:AcademicArticle type were beeing annotated.

Now I propose the function getBiboDocumentsExceptingReferences, which tries to retrieve the uri of the document that is going to be annotated. All bibo documents are iterated, and I check their URIs follow the biotea pattern while creating rdf document resources. If more than one, I select the first and I log a warning. This allows to annotate more documents in addition to bibo:AcademicArticle.